backend/requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Database
psycopg[binary]==3.2.10
psycopg-pool==3.2.3

gunicorn==21.2.0
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
python-dotenv==1.1.1
rapidfuzz==3.14.3

# Authentication (Phase 2)
PyJWT==2.8.0
bcrypt==4.1.2

# Email Service (Phase 3 - optional for Phase 2)
# sendgrid==6.11.0

# Rate Limiting (Phase 8 - optional for Phase 2)
# Flask-Limiter==3.5.0

# Google OAuth (Phase 6 - optional for Phase 2)
# google-auth==2.27.0